Recycle XLogData too
